use useFetchRoute and use it in confirm page too
